Something Something, Im still building this thing

+++ etag_resource.go
@@ -3,7 +3,7 @@ package monocle
 import "time"
 
 type EtagResource struct {
-	ID        uint      `db:"id" json:"id"`
+	ID        uint64    `db:"id" json:"id"`
 	Resource  string    `db:"resource" json:"resource"`
 	Etag      string    `db:"etag" json:"etag"`
 	Expires   time.Time `db:"expires" json:"expires"`
